Updated test_login_page
Added conditional in the setUpClass method of the TestCase as
Chrome webdriver to check whether CHROMEDRIVER_PATH env var is
set. If set, it is passed as argument to the driver, else
it is assumed that the path to the driver is set in PATH and
driver is instantiated without arguments.